pyproject.toml
[tool.poetry]
name = "dpv-tool"
version = "0.1.0"
description = ""
authors = ["Robert Bagdahn <robertbaggi@gmail.com>"]
readme = "README.md"
packages = [{include = "dpv_tool"}]

[tool.poetry.dependencies]
python = "3.11.7"
django = "^4.1.9"
celery = "^5.2.7"
environ = "^1.0"
django-environ = "^0.9.0"
python-keycloak = "^2.9.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
django-storages = "^1.13.2"
django-ebhealthcheck = "^2.0.2"
django-extensions = "^3.2.1"
django-filter = "^22.1"
mozilla-django-oidc = "^3.0.0"
django-colorfield = "^0.8.0"
django-polymorphic = "^3.1.0"
django-rest-polymorphic = "^0.1.10"
drf-api-logger = "^1.1.12"
s3transfer = "^0.6.0"
boto3 = "^1.26.59"
djangorestframework-camel-case = "^1.3.0"
openpyxl = "^3.0.10"
drf-nested-routers = "^0.93.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
